ft(schema): create a seperate folder for schema
