+++ server/middlewares/user.middlewares.js
@@ -1,39 +1,14 @@
 import Joi from 'joi';
 import joiFormater from '../helpers/joiFormater';
 import findUser from '../helpers/findUser';
+import signupSchema from '../joiSchema/signupSchema';
+import loginSchema from '../joiSchema/loginSchema';
 
 const signUpValidator = async (req, res, next) => {
-  const schema = Joi.object().keys({
-    firstname: Joi.string()
-      .min(3)
-      .max(30)
-      .required(),
-    lastname: Joi.string()
-      .min(3)
-      .max(30)
-      .required(),
-    email: Joi.string()
-      .email({ minDomainAtoms: 2 })
-      .required(),
-    password: Joi.string()
-      .min(8)
-      .alphanum()
-      .required(),
-    confirmPassword: Joi.string()
-      .required()
-      .valid(Joi.ref('password'))
-      .options({
-        language: {
-          any: {
-            allowOnly: '!!Passwords do not match',
-          },
-        },
-      }),
-  });
   const { firstname, lastname, email, password, confirmPassword } = req.body;
   const { error } = Joi.validate(
     { firstname, lastname, email, password, confirmPassword },
-    schema
+    signupSchema()
   );
   if (error) {
     const { message } = error.details[0];
@@ -45,7 +20,7 @@ const signUpValidator = async (req, res, next) => {
   const user = await findUser('email', email);
   if (user) {
     return res.status(409).send({
-      message: 'email already exist',
+      message: 'Email already exists',
     });
   }
   return next();
@@ -53,13 +28,7 @@ const signUpValidator = async (req, res, next) => {
 
 const loginValidator = (req, res, next) => {
   const { email, password } = req.body;
-  const schema = Joi.object().keys({
-    email: Joi.string()
-      .email({ minDomainAtoms: 2 })
-      .required(),
-    password: Joi.string().required(),
-  });
-  const { error } = Joi.validate({ email, password }, schema);
+  const { error } = Joi.validate({ email, password }, loginSchema());
   if (error) {
     const { message } = error.details[0];
     const formatedMessage = joiFormater(message);
